requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic-settings==2.4.0
sqlalchemy==2.0.32
alembic==1.13.2
python-multipart==0.0.9
numpy==1.26.4
opencv-python==4.9.0.80
Pillow==10.4.0
torch==2.2.2
torchvision==0.17.2
vidgear[videocapture]==0.3.2
yt-dlp>=2025.1.15
ultralytics==8.2.103
paho-mqtt==1.6.1
qrcode[pil]==8.2
firebase-admin==6.6.0